Entities that had Revelio/Accio/Depulso bugged now are unaffected by it.
Things like Minecarts would be left glowing forever or being completely unaffected by it in the first place, since they overwrite the "tick()" method without calling Entity.tick(), which is where I'm ticking down to count when to unglow it. I may be able to remake how it works so that it works on them as well.